pvr: only share scratch buffers when they're the required size
When sharing scratch buffers across framebuffers, it reallocates the scratch
buffer to ensure the smallest scratch buffer is always in use. However, the
check to do this meant that it was possible to end up with a scratch buffer that
was too small, leading to page faults or memory corruption.

Signed-off-by: Frank Binns <frank.binns@imgtec.com>
Acked-by: Erik Faye-Lund <erik.faye-lund@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/36412>