gfxstream: guest: nuke android_compat directory
With the requirement for the Android compatibility
layer lifted, we can nuke many things in "fuchsia"
directory (which was also used for guest Linux builds).

Things that might be useful later are kept, but should
probably be ported to Mesa's util layer.

Reviewed-by: Aaron Ruby <aruby@blackberry.com>
Acked-by: Yonggang Luo <luoyonggang@gmail.com>
Acked-by: Adam Jackson <ajax@redhat.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/27246>