tu: Track possible feedback loops for dynamic renderpasses
With DRLR, we unfortunately don't get notified about when there are
feedback loops. We also can't detect ahead of time the case where an
input attachment is read before written. This means we need to try and
guess in the driver when that happens based on which input attachments
we read.

There are three main things we have to do with feedback loops:

- Forcing flushing between primitives in sysmem to avoid problems with
  compressed textures before a7xx.
- Forcing late depth test if it's a depth feedback loop.
- Invalidating UCHE because we may read blit results.

The way this is implemented is currently a bit conservative, following
radv. However the impact of the flushing should be mitigated by
switching to GMEM whenever we can, which we already do.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/31261>