Revert "radv: specialize push constant stages with DGC"
This change was wrong but there is nothing testing this. For example,
if we have a pipeline with VS+GS+FS and DGC only updates push constants
for VS. On GFX9+, VS is merged to GS, so the VS push constants info
would be zero and nothing would be emitted.

This reverts commit 72c41ce594d23305eda5fc959849c778388cdb21.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/31099>