intel/perf: delete dead code.
The inner loop with p is dead, because n_passes_written is no longer
updated as of 2f2957a755c1ca98a14d8199e37c876bc45d2c8e, so it is always
comparing a uint32_t < 0, which is never true. Since the inner loop is
dead code, the pass array is dead code, as it simply keeps writing to
element 0, and but never reads or uses it, along with all of the pass
count information.

Reviewed-by: José Roberto de Souza <jose.souza@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/31213>