panvk/ci: Increase panfrost-g52-vk fraction
premerge have a time limit of 15 min, those jobs takes between 20 to 33
min with a fraction of 8 right now.

Additionally, we only have 7 VIM3 and could be trying to queue 8 jobs.

Let's fraction to 32 for now and revert back to 2 runners.

Signed-off-by: Mary Guillemard <mary.guillemard@collabora.com>
Fixes: f1ae027b9413 ("panvk/ci: Update to run full CTS on G52")
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/31404>