panfrost: add support for incremental rendering in CSF
When the tiler heap is exhausted while running an IDVS job, the
out-of-memory exception handler is gonna be executed. If that
happens, we start fragment shading for all the polygons that have
been generated until then and recycle the tiler heap memory to
continue the tiling job.

The framebuffer descriptor used for the fragment job depends on the
current pass:

  - FIRST_PASS: Used the first time the OOM handler is called. We
    want to preload/clear just like configured for the draw but
    don't want to discard the result.
  - MIDDLE_PASS: Used the subsequent times the OOM handler is
    called. We want to preload the content rendered from previous
    rendering passes and still don't want to discard the result.
  - LAST_PASS: Used for the fragment job run when back in normal
    operation. We want to preload the content from the rendering
    passes that occurred in the OOM handler and want to deal with
    the result just like it was configured initially for the draw.

If incremental rendering is not triggered by an OOM exception, the
fragment job is using the default framebuffer descriptor.

Signed-off-by: Louis-Francis Ratté-Boulianne <lfrb@collabora.com>
Signed-off-by: antonino <antonino.maniscalco@collabora.com>
Reviewed-by: Boris Brezillon <boris.brezillon@collabora.com>
Reviewed-by: Lars-Ivar Hesselberg Simonsen <lars-ivar.simonsen@arm.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/31174>