r300: always transform sin/cos input for fs
We don't need to check anymore if this was already applied and it turned
out the check was not working properly in the first place.

The check for vs is kept in place, because that one still detects that
few wine shaders already have the sin/cos input in correct range.

Signed-off-by: Pavel Ondračka <pavel.ondracka@gmail.com>
Reviewed-by: Filip Gawin <filip@gawin.net>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/32160>